1.8 alpha added dwarf(trades ai sounds and textures todo) added dwarf villages added pick,flamethrower,armor,dwarf handgun

+++ src/main/java/warhammermod/Entities/EntityHalberd.java
@@ -87,7 +87,7 @@ public class EntityHalberd extends EntityThrowable
                     }
                 }
                 if (result.entityHit != entityplayer)
-                    result.entityHit.attackEntityFrom(DamageSource.causeThrownDamage(this, this.getThrower()), (float) damage);
+                    result.entityHit.attackEntityFrom(DamageSource.causeThrownDamage(this, this.getThrower()), damage);
             }
         }
             if (!this.world.isRemote) {
